chore: move isSupportedChainName function to the utils/blockchain.ts
