+++ src/types/withdraw.ts
@@ -1,11 +1,6 @@
-import { CHAIN_IDS } from "src/constants/evm"
 import type { SendNearTransaction } from "../features/machines/publicKeyVerifierMachine"
 import type { AuthHandle } from "./authHandle"
-import type {
-  BaseTokenInfo,
-  SupportedChainName,
-  UnifiedTokenInfo,
-} from "./base"
+import type { BaseTokenInfo, UnifiedTokenInfo } from "./base"
 import type { RenderHostAppLink } from "./hostAppLink"
 import type { WalletMessage, WalletSignatureResult } from "./walletMessage"
 export type WithdrawWidgetProps = {
@@ -24,11 +19,4 @@ export type WithdrawWidgetProps = {
    * Prop is not reactive, set it once when the component is created.
    */
   referral?: string
-}
-
-export function isSupportedChainName(
-  chainName: string
-): chainName is SupportedChainName {
-  const supportedChainNames = Object.keys(CHAIN_IDS)
-  return supportedChainNames.includes(chainName)
 }
